+++ ComplimentApp/app/src/main/java/com/matikano/complimentapp/data/workers/ChildWorkerFactory.kt
@@ -1,9 +1,0 @@
-package com.matikano.complimentapp.data.workers
-
-import android.content.Context
-import androidx.work.CoroutineWorker
-import androidx.work.WorkerParameters
-
-interface ChildWorkerFactory {
-    fun create(context: Context, params: WorkerParameters): CoroutineWorker
-}