+++ ComplimentApp/app/src/main/java/com/matikano/complimentapp/di/modules/AppModule.kt
@@ -1,11 +1,13 @@
-package com.matikano.complimentapp.di
+package com.matikano.complimentapp.di.modules
 
 import android.content.Context
+import com.matikano.complimentapp.data.local.NotificationPreferences
 import com.matikano.complimentapp.data.remote.ComplimentApi
 import com.matikano.complimentapp.util.Constants.API.BASE_URL
 import dagger.Module
 import dagger.Provides
 import dagger.hilt.InstallIn
+import dagger.hilt.android.qualifiers.ApplicationContext
 import dagger.hilt.components.SingletonComponent
 import retrofit2.Retrofit
 import retrofit2.converter.moshi.MoshiConverterFactory
@@ -25,6 +27,11 @@ object AppModule {
             .create(ComplimentApi::class.java)
 
 
+    @Provides
+    @Singleton
+    fun provideNotificationPreferences(
+        @ApplicationContext context: Context
+    ): NotificationPreferences = NotificationPreferences(context)
 
 
 }