+++ ComplimentApp/app/src/main/java/com/matikano/complimentapp/presentation/compliment/ComplimentEvent.kt
@@ -1,6 +1,6 @@
 package com.matikano.complimentapp.presentation.compliment
 
 sealed class ComplimentEvent {
-    data class OnLoadCompliment(val content: String?): ComplimentEvent()
+    data class OnLoadCompliment(val content: String? = null): ComplimentEvent()
     object OnRefresh: ComplimentEvent()
 }