+++ ComplimentApp/CHANGELOG.md
@@ -5,6 +5,17 @@ All notable changes to this project will be documented in this file.
 The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/),
 and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
 
+## [0.2.0] - 2022-10-17
+
+### Added
+* HiltBroadcastReceiver abstract class for creating implementations of BroadcastReceiver class that can have injected dependencies
+* NotificationAlarmReceiver class that handles broadcast sent by AlarmManager - displays notifications with daily compliments
+
+### Changed
+* Code cleanup
+
+### Deleted
+* WorkManager classes and implementation of periodic task execution since it is nor reliable and does not work periodically when device is in Doze mode and Power Safe mode
 
 ## [0.1.1] - 2022-10-17